Conflicts:
	sdklib/src/main/java/com/android/sdklib/BuildToolInfo.java

+++ build-system/gradle/src/main/groovy/com/android/build/gradle/tasks/PreDex.groovy
@@ -22,12 +22,13 @@ import com.android.builder.core.AndroidBuilder
 import com.android.builder.core.DexOptions
 import com.android.ide.common.internal.WaitableExecutor
 import com.google.common.base.Charsets
-import com.google.common.collect.ImmutableSet
+import com.google.common.collect.Lists
 import com.google.common.collect.Sets
 import com.google.common.hash.HashCode
 import com.google.common.hash.HashFunction
 import com.google.common.hash.Hashing
 import com.google.common.io.Files
+import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.InputFiles
 import org.gradle.api.tasks.Nested
 import org.gradle.api.tasks.OutputDirectory
@@ -41,6 +42,10 @@ public class PreDex extends BaseTask {
     // ----- PUBLIC TASK API -----
 
     // ----- PRIVATE TASK API -----
+    @Input
+    String getBuildToolsVersion() {
+        plugin.extension.buildToolsRevision
+    }
 
     // this is used automatically by Gradle, even though nothing
     // in the class uses it.
@@ -54,49 +59,70 @@ public class PreDex extends BaseTask {
     @Nested
     DexOptionsImpl dexOptions
 
+    @Input
+    boolean multiDex
+
     @TaskAction
     void taskAction(IncrementalTaskInputs taskInputs) {
+
+        final boolean multiDexEnabled = getMultiDex()
+
         final File outFolder = getOutputFolder()
 
+        boolean incremental = taskInputs.isIncremental()
         // if we are not in incremental mode, then outOfDate will contain
-        // all th files, but first we need to delete the previous output
-        if (!taskInputs.isIncremental()) {
+        // all the files, but first we need to delete the previous output
+        if (!incremental) {
             emptyFolder(outFolder)
         }
 
         final Set<String> hashs = Sets.newHashSet()
         final WaitableExecutor<Void> executor = new WaitableExecutor<Void>()
-        final ImmutableSet.Builder<File> inputFileDetailses = ImmutableSet.builder()
+        final List<File> inputFileDetails = Lists.newArrayList()
 
         taskInputs.outOfDate { final change ->
-            inputFileDetailses.add(change.file)
+            inputFileDetails.add(change.file)
         }
 
-        for (final File file : inputFileDetailses.build()) {
-            Callable<Void> action = new PreDexTask(file, hashs);
+        for (final File file : inputFileDetails) {
+            Callable<Void> action = new PreDexTask(outFolder, file, hashs,
+                    multiDexEnabled);
             executor.execute(action);
         }
 
-        taskInputs.removed { change ->
-            //noinspection GroovyAssignabilityCheck
-            File preDexedFile = getDexFileName(outFolder, change.file)
-            preDexedFile.delete()
+        if (incremental) {
+            taskInputs.removed { change ->
+                //noinspection GroovyAssignabilityCheck
+                File preDexedFile = getDexFileName(outFolder, change.file)
+                if (preDexedFile.isDirectory()) {
+                    preDexedFile.deleteDir()
+                } else {
+                    preDexedFile.delete()
+                }
+            }
         }
 
         executor.waitForTasksWithQuickFail(false)
     }
 
     private final class PreDexTask implements Callable<Void> {
-        private final File fileToProcess;
-        private final Set<String> hashs;
+        private final File outFolder
+        private final File fileToProcess
+        private final Set<String> hashs
+        private final boolean multiDexEnabled
         private final DexOptions options = getDexOptions()
         private final AndroidBuilder builder = getBuilder()
-        final File outFolder = getOutputFolder()
 
 
-        private PreDexTask(File file, Set<String> hashs) {
-            this.fileToProcess = file;
-            this.hashs = hashs;
+        private PreDexTask(
+                File outFolder,
+                File file,
+                Set<String> hashs,
+                boolean multiDexEnabled) {
+            this.outFolder = outFolder
+            this.fileToProcess = file
+            this.hashs = hashs
+            this.multiDexEnabled = multiDexEnabled
         }
 
         @Override
@@ -114,8 +140,13 @@ public class PreDex extends BaseTask {
 
             //noinspection GroovyAssignabilityCheck
             File preDexedFile = getDexFileName(outFolder, fileToProcess)
+
+            if (multiDexEnabled) {
+                preDexedFile.mkdirs()
+            }
+
             //noinspection GroovyAssignabilityCheck
-            builder.preDexLibrary(fileToProcess, preDexedFile, options)
+            builder.preDexLibrary(fileToProcess, preDexedFile, multiDexEnabled, options)
 
             return null
         }
@@ -136,8 +167,11 @@ public class PreDex extends BaseTask {
      * if there are 2 libraries with the same file names (but different
      * paths)
      *
+     * If multidex is enabled the return File is actually a folder.
+     *
      * @param outFolder the output folder.
      * @param inputFile the library
+     * @param multidex whether multidex is enabled.
      * @return
      */
     @NonNull
